+++ lib/ocr/scan.py
